Move files to the same directory
