+++ src/get-buffer-response-browser.ts
