+++ src/index.ts
@@ -1,9 +1,9 @@
 import { endpoint } from '@octokit/endpoint'
 import getUserAgent from 'universal-user-agent'
 
-import { version } from './package.json';
+import { version } from '../package.json';
 const userAgent = `octokit-request.js/${version} ${getUserAgent()}`
-import withDefaults from './lib/with-defaults'
+import withDefaults from './with-defaults'
 
 export default withDefaults(endpoint, {
   headers: {
