+++ tsconfig.json
@@ -15,7 +15,7 @@
         "declarationMap": true, /* Generates a sourcemap for each corresponding '.d.ts' file. */
         "sourceMap": true, /* Generates corresponding '.map' file. */
         // "outFile": "./",                       /* Concatenate and emit output to single file. */
-        "outDir": "./lib", /* Redirect output structure to the directory. */
+        "outDir": "src", /* Redirect output structure to the directory. */
         // "rootDir": "./",                       /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
         "composite": true, /* Enable project compilation */
         // "removeComments": true,                /* Do not emit comments to output. */
@@ -57,8 +57,7 @@
         // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     },
     "include": [
-        "lib/*.ts",
-        "index.ts",
+        "src/*",
         "universal-user-agent.d.ts"
     ],
     "exclude": []
